tests/wpt/tests/css/css-lists/parsing/list-style-shorthand.sub.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Lists: list-style sets longhands</title>
<link rel="help" href="https://drafts.csswg.org/css-lists-3/#propdef-list-style">
<meta name="assert" content="list-style supports the full grammar '<'list-style-position'> || <'list-style-image'> || <'list-style-type'>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
<script>
test_shorthand_value('list-style', 'square url("https://{{host}}/") inside', {
  'list-style-position': 'inside',
  'list-style-image': 'url("https://{{host}}/")',
  'list-style-type': 'square'
});
</script>
</body>
</html>
